// src/app/typescript-angular-client-generated/model/inspeccionTecnicaViewDto.ts
/**
 * Valoriza Park API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v2
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface InspeccionTecnicaViewDto { 
    id?: number;
    matricula?: string;
    claveGuid?: string;
    fecha?: string;
    fechaProxima?: string;
    comentarios?: string;
    createdBy?: string;
    createdAt?: Date;
    changedBy?: string;
    changedAt?: Date;
    centro?: string;
    delegacion?: string;
    familia?: string;
    subfamilia?: string;
}